moved indices from hydro to indices.
